Version 3.0/requirements.txt
openpyxl==3.1.5
pandas==2.2.3
prettytable==3.11.0
py==1.11.0
PyYAML==6.0.2
SQLAlchemy==2.0.36
streamlit==1.39.0
streamlit_authenticator==0.4.1
